refactor: migrate everything to use hooks
Migrates the existing code base to use hooks. This removes
some of the dependencies from the bundle (prop-types, redux and
react-redux modules that will be tree-shook, etc) as most of them
are now simple function components.

Also migrated the testing from enzyme to react-testing-library,
and migrated the chart component's testing to be an integration
test.