feat: re-organize directory structure
* Remove suffix from each component's names, as it's not
necessary to have them.

* Rename directory "containers" to "pages". It feels more
semantic to name these components as pages, as how they are being
routed makes them simulate different website pages.

* Move Login and Logout from "container" to "components".
The reasoning behind this is that components can be thought of
as components that persist between multiple website pages.
As both of these components are exactly that, it makes sense for
them to be there.